refactor(render): cut over world renderer into package modules
