+++ src/crimson/render/world/__init__.py
@@ -1,0 +1,6 @@
+from __future__ import annotations
+
+from .constants import monster_vision_fade_alpha
+from .renderer import WorldRenderer
+
+__all__ = ["WorldRenderer", "monster_vision_fade_alpha"]
