+++ src/crimson/render/world/mixin_base.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+from typing import Any
+
+
+class WorldRendererMixinBase:
+    """Allow cross-mixin attribute access without duplicating giant protocols."""
+
+    def __getattr__(self, name: str) -> Any:
+        raise AttributeError(name)
