+++ tests/test_monster_vision_perk.py
@@ -7,7 +7,7 @@ import math
 from crimson.sim.state_types import PlayerState
 from crimson.perks import PerkId
 from crimson.perks.helpers import perk_active
-from crimson.render.world_renderer import monster_vision_fade_alpha
+from crimson.render.world import monster_vision_fade_alpha
 
 
 def test_monster_vision_fade_alpha_matches_death_stage_clamp() -> None:
