+++ tests/test_world_renderer_projectile_trail.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from crimson.projectiles import ProjectileTypeId
-from crimson.render.world_renderer import WorldRenderer
+from crimson.render.world import WorldRenderer
 from grim.geom import Vec2
 
 
@@ -17,17 +17,17 @@ class _WorldStub:
 def test_draw_bullet_trail_zero_length_still_counts_as_drawn(monkeypatch) -> None:
     vertices: list[tuple[float, float]] = []
 
-    monkeypatch.setattr("crimson.render.world_renderer.rl.begin_blend_mode", lambda _mode: None)
-    monkeypatch.setattr("crimson.render.world_renderer.rl.rl_set_texture", lambda _tex_id: None)
-    monkeypatch.setattr("crimson.render.world_renderer.rl.rl_begin", lambda _mode: None)
-    monkeypatch.setattr("crimson.render.world_renderer.rl.rl_color4ub", lambda _r, _g, _b, _a: None)
-    monkeypatch.setattr("crimson.render.world_renderer.rl.rl_tex_coord2f", lambda _u, _v: None)
+    monkeypatch.setattr("crimson.render.world.projectiles.rl.begin_blend_mode", lambda _mode: None)
+    monkeypatch.setattr("crimson.render.world.projectiles.rl.rl_set_texture", lambda _tex_id: None)
+    monkeypatch.setattr("crimson.render.world.projectiles.rl.rl_begin", lambda _mode: None)
+    monkeypatch.setattr("crimson.render.world.projectiles.rl.rl_color4ub", lambda _r, _g, _b, _a: None)
+    monkeypatch.setattr("crimson.render.world.projectiles.rl.rl_tex_coord2f", lambda _u, _v: None)
     monkeypatch.setattr(
-        "crimson.render.world_renderer.rl.rl_vertex2f",
+        "crimson.render.world.projectiles.rl.rl_vertex2f",
         lambda x, y: vertices.append((float(x), float(y))),
     )
-    monkeypatch.setattr("crimson.render.world_renderer.rl.rl_end", lambda: None)
-    monkeypatch.setattr("crimson.render.world_renderer.rl.end_blend_mode", lambda: None)
+    monkeypatch.setattr("crimson.render.world.projectiles.rl.rl_end", lambda: None)
+    monkeypatch.setattr("crimson.render.world.projectiles.rl.end_blend_mode", lambda: None)
 
     renderer = WorldRenderer(_world=_WorldStub())  # type: ignore[arg-type]
